+++ R/wrangling_dataset_issues.R
@@ -101,69 +101,10 @@ get_issues <- function(
     state <- match.arg(state)
 
     if (source == "online") {
-        info_owner <- try(expr = {
-            gh::gh(
-                endpoint = "/users/:owner",
-                owner = owner,
-                .limit = Inf
-            )
-        })
-        if (inherits(info_owner, "try-error")) {
-            message_response <- attr(
-                info_owner,
-                "condition"
-            )$response_content$message
-            if (message_response == "Not Found") {
-                stop(
-                    owner,
-                    " is not a valid GitHub user.\n",
-                    "The argument owner must be a valid GitHub user.",
-                    call. = FALSE
-                )
-            } else if (grepl("API rate limit exceeded", message_response,
-                             fixed = TRUE)) {
-                warning(
-                    message_response,
-                    "\n",
-                    attr(info_owner, "condition")$body,
-                    call. = FALSE
-                )
-                return(new_issues())
-            } else {
-                stop("Weird message... Contact the maintainer of the package.",
-                     call. = FALSE)
-            }
-        }
 
-        owner_type <- info_owner$type
         if (is.null(repo)) {
             if (verbose) cat("Try to find all repositories...")
-
-            if (owner_type == "User") {
-                endpoint <- "/users/:owner/repos"
-            } else if (owner_type == "Organization") {
-                endpoint <- "/orgs/:owner/repos"
-            } else {
-                stop("owner type not taken into account", call. = FALSE)
-            }
-
-            list_public_repo <- gh::gh(
-                endpoint = endpoint,
-                owner = owner,
-                .limit = Inf
-            ) |>
-                vapply(FUN = "[[", "name", FUN.VALUE = character(1L))
-
-            list_private_repo <- gh::gh(
-                endpoint = "/user/repos",
-                .limit = Inf,
-                visibility = "private"
-            ) |>
-                Filter(f = \(.x) .x$owner$login == owner) |>
-                vapply(FUN = "[[", "name", FUN.VALUE = character(1L))
-
-            list_repo <- unique(c(list_private_repo, list_public_repo))
-
+            list_repo <- get_all_repos(owner)
             if (verbose) cat(" Done!\n")
 
             issues <- lapply(
@@ -199,10 +140,7 @@ get_issues <- function(
             if (message_response == "Not Found") {
                 stop(
                     repo,
-                    " is not a",
-                    ifelse(owner_type == "Organization", "n ", " "),
-                    tolower(owner_type),
-                    " repository from ",
+                    " is not a repository from ",
                     owner,
                     ".\nThe argument repo must be a valid GitHub repo name.",
                     call. = FALSE
@@ -367,8 +305,8 @@ format_issues <- function(
         "^https://api.github.com/repos/([^/]+)/([^/]+)/issues/\\d+$",
         urls,
         proto = data.frame(
-            owners = character(),
-            repos = character(),
+            owner = character(),
+            repo = character(),
             stringsAsFactors = FALSE
         )
     )
@@ -459,8 +397,8 @@ format_issues <- function(
             },
             FUN.VALUE = character(1L)
         ),
-        owner = structurel$owners,
-        repo = structurel$repos
+        owner = structurel$owner,
+        repo = structurel$repo
     )
 
     return(issues)